PSTRESS-160: Divide `mysql_option_tester.py` into `mysql_option_tester.py` and `mysql_utils.py`
